+++ js-client/CHANGELOG.md
@@ -2,11 +2,13 @@
 
 ## 0.3.1
 
-- When running a step outside of a run, automatically wrap it in a run
+- When running a step outside of a run, automatically wrap it in a run.
 - Allow disabling all event logging in an application.
-- Allow explicitly passing a run context to `runStep`, in case it can not be retrieved from the normal AsyncLocalStorage context.
+- Allow explicitly passing a run context to `runStep`, in case it can not be retrieved from the normal `AsyncLocalStorage` context.
 - The client is now also an `EventEmitter`, and will emit any events passed to `client.event()`.
-- Add a `withMetadata` function to the client, which returns a child client with updated default metadata values in the requests. This client shares the same EventEmitter and other settings with its parent.
+- Add a `withMetadata` function to the client, which returns a child client with updated default metadata values in the requests. This client shares the same `EventEmitter` and other settings with its parent.
+- Allow passing a custom Chronicle client instance to `runStep`.
+- Allow passing metadata to `startRun` or `runStep`, which will automatically create a child Chronicle client using `withMetadata`.
 
 ## 0.3.0